remove dupe workspace file and method
